% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{c_to_k}
\alias{c_to_k}
\title{Convert Celsius to Kelvin}
\usage{
c_to_k(c)
}
\arguments{
\item{c}{Numeric or SpatRaster in Celsius.}
}
\value{
Same type, in Kelvin.
}
\description{
Convert Celsius to Kelvin
}
