% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_airtemp.R
\name{.grid_cells_as_sf}
\alias{.grid_cells_as_sf}
\title{Build an sf POINT layer of every raster cell centre, in a target CRS}
\usage{
.grid_cells_as_sf(template, target_crs)
}
\description{
Build an sf POINT layer of every raster cell centre, in a target CRS
}
\keyword{internal}
