% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{k_to_c}
\alias{k_to_c}
\title{Convert Kelvin to Celsius}
\usage{
k_to_c(k)
}
\arguments{
\item{k}{Numeric or SpatRaster in Kelvin.}
}
\value{
Same type, in Celsius.
}
\description{
Convert Kelvin to Celsius
}
