% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst_appeears.R
\name{lst_appeears_check_status}
\alias{lst_appeears_check_status}
\title{Check status of an AppEEARS task}
\usage{
lst_appeears_check_status(task_id, token)
}
\arguments{
\item{task_id}{Character task ID from \code{\link{lst_appeears_submit_task}}.}

\item{token}{Bearer token.}
}
\value{
Character status: one of \code{"pending"}, \code{"processing"},
  \code{"done"}, \code{"error"}.
}
\description{
Check status of an AppEEARS task
}
