% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst_appeears.R
\name{lst_appeears_download}
\alias{lst_appeears_download}
\title{Download completed AppEEARS task output}
\usage{
lst_appeears_download(task_id, token, outdir)
}
\arguments{
\item{task_id}{Character task ID.}

\item{token}{Bearer token.}

\item{outdir}{Directory to write files to. Created if it doesn't exist.}
}
\value{
Character vector of downloaded file paths, invisibly.
}
\description{
Downloads all GeoTIFF bundle files for a completed task to a local
directory.
}
