% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst_local.R
\name{mask_lst_by_qc}
\alias{mask_lst_by_qc}
\title{Mask an LST raster using a QC raster}
\usage{
mask_lst_by_qc(lst, qc, keep_classes = c("good", "marginal"))
}
\arguments{
\item{lst}{A LST \code{SpatRaster}.}

\item{qc}{A raw QC \code{SpatRaster}, same extent/resolution as \code{lst}.}

\item{keep_classes}{See \code{\link{decode_modis_qc}}.}
}
\value{
Masked \code{SpatRaster}.
}
\description{
Mask an LST raster using a QC raster
}
