% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst_local.R
\name{mosaic_lst}
\alias{mosaic_lst}
\title{Mosaic multiple LST tiles/dates}
\usage{
mosaic_lst(rast_list, fun = "mean")
}
\arguments{
\item{rast_list}{A list of \code{SpatRaster} objects, or a character
vector of file paths, to mosaic together.}

\item{fun}{Function used to combine overlapping cells. Default \code{"mean"}.}
}
\value{
A single mosaicked \code{SpatRaster}.
}
\description{
Mosaic multiple LST tiles/dates
}
