% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_hotspots}
\alias{plot_hotspots}
\title{Quick diagnostic plot of a hotspot classification raster}
\usage{
plot_hotspots(class_rast, title = "Hotspot analysis (Gi*)")
}
\arguments{
\item{class_rast}{Output of \code{\link{classify_hotspots}}.}

\item{title}{Plot title.}
}
\value{
Invisibly, the plotted \code{SpatRaster}.
}
\description{
Quick diagnostic plot of a hotspot classification raster
}
