% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\name{read_landsat_sr}
\alias{read_landsat_sr}
\title{Leggi una banda di riflettanza superficiale Landsat Collection 2 Livello 2}
\usage{
read_landsat_sr(path)
}
\arguments{
\item{path}{Percorso del file GeoTIFF della banda SR (es. SR_B4, SR_B5, SR_B6).}
}
\value{
Un \code{SpatRaster} con valori di riflettanza fisica (0-1 circa).
}
\description{
Applica i fattori di scala ufficiali per le bande SR (Surface Reflectance)
di Landsat Collection 2 Livello 2 (diversi da quelli della banda termica
ST_B10).
}
